debugging
# ISSUE
the code is supposed to load a termID from the Intent
when the default value is set to 0 on the getIntent.getIntExtra nothing appears
when the default value is set to 1 items appear. Value isn't being loaded somewhere

# CAUSES
maybe I'm not passing the intent on click?
CourseAdapter has the intent?
	no, TermActivity has the intent that passes to the CourseList

# SUMMARY
- adding debug logs to figure out the issue
- checking Course adapter, termdetails activity, terms actiivty, and coursesActivity
- found that TermDetailsActivity is the one utilizing the TermAdapter intent that can be passed to the CourseActivity
- FIXED

# SOLUTION
TermDetailsActivity: added putExtra on the intent method to resolve the missing intent data in CourseActivity